archive/register/pyproject.toml
[project]
name = "vecgraph-main"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "chromadb==0.5.0",
    "couchdb==1.2",
    "fastapi==0.115.0",
    "jwt>=1.4.0",
    "nebula3-python>=3.8.3",
    "neo4j==5.21.0",
    "numpy<2.0.0",
    "prometheus-client==0.20.0",
    "pydantic>=2.12.4",
    "pytest>=9.0.1",
    "python-dotenv==1.0.1",
    "python-multipart>=0.0.20",
    "pyyaml>=6.0.3",
    "reportlab>=4.4.5",
    "requests==2.32.3",
    "sentence-transformers==3.0.1",
    "uvicorn[standard]==0.30.0",
    "weaviate-client>=4.18.0",
]
